Complete Github Cheaters Checking
Added a Cheaters.Java which is a standalone program (but using Problem.java) to process downloaded GitHub repositories.  

I've included a cheaters.txt with all problems from Ramakastriot and Creede15.

If you want to process more Github Repositories unzip the Github Cheaters into a directory **above** your source for it to read it.  This is to prevent Eclipse from compiling all of them.  Cheaters.java will traverse the directory for all repositories, use the filenames as problem names, match them up with BJP Problem Names.csv, and generate a cheaters.txt file.  I had to fix several filenames to match the BJP names.  There are also several old problems that were in BJP3 but no longer in BJP 4.